apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "preguntalo-api"
version = "0.1.0"
description = "FastAPI backend for multilingual manual search"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "openai>=1.68.2",
  "pgvector>=0.3.6",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "python-docx>=1.1.2",
  "python-multipart>=0.0.9",
  "pypdf>=5.4.0",
  "redis>=5.0.8",
  "boto3>=1.35.0",
  "httpx>=0.27.0"
]

[tool.setuptools]
packages = ["app"]
